Fix references for renamed files
