+++ src/components/Menu.js
@@ -2,8 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Icon } from 'network-canvas-ui';
 import { debounce } from 'lodash';
-import { Scroller } from '.';
-import { MenuItem } from './Elements';
+import { MenuItem, Scroller } from './';
 
 const closeEvents = [
   'click',
