+++ src/components/__tests__/Pips.test.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import { shallow } from 'enzyme';
-import Pips from '../../Elements/Pips';
+import Pips from '../../Pips';
 
 describe('Pips component', () => {
   it('renders pips matching count', () => {
