+++ src/components/__tests__/Prompt.test.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import { shallow } from 'enzyme';
-import Prompt from '../../Elements/Prompt';
+import Prompt from '../../Prompt';
 
 describe('Prompt component', () => {
   it('renders prompt with title', () => {
