+++ src/containers/Form.js
@@ -4,9 +4,9 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { reduxForm, getFormValues, getFormMeta } from 'redux-form';
 import { Button } from 'network-canvas-ui';
-import { autoInitialisedForm } from '../../behaviours';
-import { Field } from '../../containers/Elements';
-import { makeRehydrateFields } from '../../selectors/rehydrate';
+import { autoInitialisedForm } from '../behaviours';
+import { Field } from '../containers/';
+import { makeRehydrateFields } from '../selectors/rehydrate';
 
 /**
   * Renders a redux form that contains fields according to a `fields` config.
