+++ src/containers/FormWizard.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Icon } from 'network-canvas-ui';
-import { Pips } from '../../components/Elements';
-import { Form } from '.';
+import { Pips } from '../components/';
+import { Form } from '../containers/';
 
 class FormWizard extends Component {
   constructor(props) {
