+++ src/containers/Interfaces/index.js
@@ -1,4 +1,4 @@
 /* eslint-disable import/prefer-default-export */
-export { default as NameGeneratorInterface } from './NameGeneratorInterface';
-export { default as SociogramInterface } from './SociogramInterface';
-export { default as QuizInterface } from './QuizInterface';
+export { default as NameGeneratorInterface } from './NameGenerator';
+export { default as SociogramInterface } from './Sociogram';
+export { default as QuizInterface } from './Quiz';
