+++ src/containers/NodeBucket.js
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { compose } from 'recompose';
 import { Node } from 'network-canvas-ui';
-import { makeGetNextUnplacedNode, makeGetSociogramOptions } from '../../selectors/sociogram';
-import { DragSource } from '../../behaviours/DragAndDrop';
+import { makeGetNextUnplacedNode, makeGetSociogramOptions } from '../selectors/sociogram';
+import { DragSource } from '../behaviours/DragAndDrop';
 
 const EnhancedNode = DragSource(Node);
 const label = node => node.nickname;
