+++ src/containers/NodePanels.js
@@ -4,13 +4,13 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { colorDictionary } from 'network-canvas-ui';
 import { includes, map, differenceBy } from 'lodash';
-import { networkNodes, makeNetworkNodesForOtherPrompts } from '../../selectors/interface';
-import { getExternalData } from '../../selectors/protocol';
-import { actionCreators as networkActions } from '../../ducks/modules/network';
-import { makeGetPromptNodeAttributes } from '../../selectors/name-generator';
-import { Panel, Panels, NodeList } from '../../components/Elements';
-import { MonitorDragSource } from '../../behaviours/DragAndDrop';
-import configurePanels from '../../behaviours/configurePanels';
+import { networkNodes, makeNetworkNodesForOtherPrompts } from '../selectors/interface';
+import { getExternalData } from '../selectors/protocol';
+import { actionCreators as networkActions } from '../ducks/modules/network';
+import { makeGetPromptNodeAttributes } from '../selectors/name-generator';
+import { Panel, Panels, NodeList } from '../components/';
+import { MonitorDragSource } from '../behaviours/DragAndDrop';
+import configurePanels from '../behaviours/configurePanels';
 
 const colorPresets = [
   colorDictionary['edge-alt-1'],
